chore: cleanup code organization
