+++ crates/holochain-manager-uniffi/src/lib.rs
@@ -1,96 +1,5 @@
 uniffi::setup_scaffolding!();
 
-use holochain_manager::{HolochainRuntime, HolochainManagerConfig, launch::launch_holochain_runtime, utils::vec_to_locked, WANNetworkConfig};
-use std::convert::Infallible;
-use std::path::PathBuf;
-use std::str::FromStr;
-use std::sync::Mutex;
-use std::sync::Arc;
-use std::sync::PoisonError;
-use url2::{Url2, Url2Error};
-
-#[derive(uniffi::Error, thiserror::Error, Debug)]
-pub enum HolochainRuntimeFFIError {
-    #[error("Holochain Manager Error: {0}")]
-    HolochainManagerError(String),
-
-    #[error(transparent)]
-    HolochainRuntimeFFIConfigError(#[from] HolochainRuntimeFFIConfigError),
-
-    #[error("Mutex poisoned")]
-    PoisonError,
-
-    #[error("IO Error: {0}")]
-    IOError(String)
-}
-impl<T> From<PoisonError<T>> for HolochainRuntimeFFIError {
-    fn from(_err: PoisonError<T>) -> Self {
-        Self::PoisonError
-    }
-}
-
-
-#[derive(uniffi::Error, thiserror::Error, Debug)]
-#[uniffi(flat_error)]
-pub enum HolochainRuntimeFFIConfigError {
-    #[error(transparent)]
-    Url2Error(#[from] Url2Error),
-
-    #[error(transparent)]
-    Infallible(#[from] Infallible),
-}
-
-#[derive(uniffi::Record)]
-pub struct HolochainRuntimeFFIConfig {
-    /// URL of bootstrap server
-    bootstrap_url: String,
-
-    /// URL of signal server
-    signal_url: String,
-
-    /// Path to directory where conductor data will be stored
-    holochain_dir: String,
-}
-
-impl TryInto<HolochainManagerConfig> for HolochainRuntimeFFIConfig {
-    type Error = HolochainRuntimeFFIConfigError;
-    fn try_into(self) -> Result<HolochainManagerConfig, Self::Error> {
-        Ok(HolochainManagerConfig {
-            wan_network_config: Some(WANNetworkConfig {
-                bootstrap_url: Url2::try_parse(self.bootstrap_url)?,
-                signal_url: Url2::try_parse(self.signal_url)?,
-            }),
-            holochain_dir: PathBuf::from_str(self.holochain_dir.as_str())?,
-        })
-    }
-}
-
-#[derive(uniffi::Object)]
-pub struct HolochainRuntimeFFI {
-    runtime: Arc<Mutex<Option<HolochainRuntime>>>,
-}
-
-#[uniffi::export]
-impl HolochainRuntimeFFI {
-    #[uniffi::constructor]
-    pub fn new() -> Self {
-        Self {
-            runtime: Arc::new(Mutex::new(None)),
-        }
-    }
-
-    pub async fn launch(&self, passphrase: Vec<u8>, config: HolochainRuntimeFFIConfig) -> Result<(), HolochainRuntimeFFIError> {
-        let runtime = launch_holochain_runtime(
-            vec_to_locked(passphrase).map_err(|e| HolochainRuntimeFFIError::IOError(e.to_string()))?,
-            config.try_into()?
-        )
-        .await
-        .map_err(|e| HolochainRuntimeFFIError::HolochainManagerError(e.to_string()))?;
-    
-        let mut runtime_lock = self.runtime.lock()?;
-        *runtime_lock = Some(runtime);
-
-        Ok(())
-    }
-
-}
+mod error;
+mod runtime;
+pub use runtime::*;
